Moved conway_voting, implementing unit tests
