+++ modules/governance_state/src/governance_state.rs
@@ -22,6 +22,8 @@ use tracing::{error, info, info_span, Instrument};
 mod alonzo_babbage_voting;
 mod state;
 mod voting_state;
+mod conway_voting;
+
 use state::State;
 use voting_state::VotingRegistrationState;
