+++ modules/governance_state/src/state.rs
@@ -17,58 +17,21 @@ use anyhow::{anyhow, bail, Result};
 use caryatid_sdk::Context;
 use hex::ToHex;
 use std::{collections::{HashMap, HashSet}, sync::Arc};
-use std::fs::OpenOptions;
-use std::io::Write;
 use tracing::{debug, error, info};
-
-#[derive(Debug)]
-pub struct ActionStatus {
-    expiration_epoch: u64,
-    ratification_epoch: Option<u64>,
-    enactment_epoch: Option<u64>
-}
+use crate::conway_voting::ConwayVoting;
 
 pub struct State {
     pub enact_state_topic: String,
-    pub verification_output_file: Option<String>,
     pub context: Arc<Context<Message>>,
 
-    proposal_count: usize,
-
-    pub action_proposal_count: usize,
-    pub votes_count: usize,
     pub drep_stake_messages_count: usize,
 
     current_era: Era,
-    conway: Option<ConwayParams>,
     drep_stake: HashMap<DRepCredential, Lovelace>,
     spo_stake: HashMap<KeyHash, DelegatedStake>,
 
     alonzo_babbage_voting: AlonzoBabbageVoting,
-    proposals: HashMap<GovActionId, (u64, ProposalProcedure)>,
-    votes: HashMap<GovActionId, HashMap<Voter, (TxHash, VotingProcedure)>>,
-    action_status: HashMap<GovActionId, ActionStatus>
-}
-
-impl ActionStatus {
-    pub fn new(current_epoch: u64, voting_length: u64) -> Self {
-        Self {
-            expiration_epoch: current_epoch + voting_length,
-            ratification_epoch: None,
-            enactment_epoch: None
-        }
-    }
-
-    pub fn is_accepted(&self) -> bool {
-        self.ratification_epoch.is_some()
-    }
-
-    pub fn is_enacted(&self, epoch: u64) -> bool {
-        match self.enactment_epoch {
-            None => false,
-            Some(enact_epoch) => enact_epoch <= epoch,
-        }
-    }
+    conway_voting: ConwayVoting
 }
 
 impl State {
@@ -81,22 +44,15 @@ impl State {
             context,
             enact_state_topic,
 
-            proposal_count: 0,
-            action_proposal_count: 0,
-            votes_count: 0,
             drep_stake_messages_count: 0,
 
-            conway: None,
             current_era: Era::default(),
 
             alonzo_babbage_voting: AlonzoBabbageVoting::default(),
-            proposals: HashMap::new(),
-            votes: HashMap::new(),
-            action_status: HashMap::new(),
+            conway_voting: ConwayVoting::new(verification_output_file),
 
             drep_stake: HashMap::new(),
             spo_stake: HashMap::new(),
-            verification_output_file,
         }
     }
 
@@ -119,7 +75,7 @@ impl State {
             self.alonzo_babbage_voting.update_parameters(p.epoch_length, p.update_quorum);
         }
         if message.params.conway.is_some() {
-            self.conway = message.params.conway.clone();
+            self.conway_voting.update_parameters(&message.params.conway);
         }
 
         Ok(())
@@ -162,15 +118,14 @@ impl State {
         } else {
             // Conway governance
             for pproc in &governance_message.proposal_procedures {
-                self.proposal_count += 1;
-                if let Err(e) = self.insert_proposal_procedure(block.epoch, pproc) {
+                if let Err(e) = self.conway_voting.insert_proposal_procedure(block.epoch, pproc) {
                     error!("Error handling governance_message: '{}'", e);
                 }
             }
 
             for (trans, vproc) in &governance_message.voting_procedures {
                 for (voter, voter_votes) in vproc.votes.iter() {
-                    if let Err(e) = self.insert_voting_procedure(voter, trans, voter_votes) {
+                    if let Err(e) = self.conway_voting.insert_voting_procedure(voter, trans, voter_votes) {
                         error!(
                             "Error handling governance voting block {}, trans {}: '{}'",
                             block.number,
@@ -178,327 +133,19 @@ impl State {
                             e
                         );
                     }
-                    self.votes_count += voter_votes.voting_procedures.len();
                 }
             }
         }
 
         Ok(())
-    }
-
-    pub fn get_conway_params(&self) -> Result<&ConwayParams> {
-        self.conway.as_ref().ok_or_else(|| anyhow!("Conway parameters not available"))
-    }
-
-    #[allow(dead_code)]
-    fn have_committee(&self) -> bool {
-        !self.conway.iter().any(|c| c.committee.is_empty())
     }
 
     /// Update proposals memory cache
-    fn insert_proposal_procedure(&mut self, epoch: u64, proc: &ProposalProcedure) -> Result<()> {
-        self.action_proposal_count += 1;
-        let prev = self.proposals.insert(proc.gov_action_id.clone(), (epoch, proc.clone()));
-        if let Some(prev) = prev {
-            return Err(anyhow!(
-                "Governance procedure {} already exists! New: {:?}, old: {:?}",
-                proc.gov_action_id,
-                (epoch, proc),
-                prev
-            ));
-        }
-
-        let prev = self.action_status.insert(
-            proc.gov_action_id.clone(),
-            ActionStatus::new(epoch, self.get_conway_params()?.gov_action_lifetime as u64)
-        );
-        if let Some(prev) = prev {
-            return Err(anyhow!(
-                "Governance procedure {} action status already exists! Old: {:?}",
-                proc.gov_action_id,
-                prev
-            ));
-        }
-
-        Ok(())
-    }
-
-    /// Update votes memory cache
-    fn insert_voting_procedure(
-        &mut self,
-        voter: &Voter,
-        transaction: &TxHash,
-        voter_votes: &SingleVoterVotes,
-    ) -> Result<()> {
-        for (action_id, procedure) in voter_votes.voting_procedures.iter() {
-            let votes = self.votes.entry(action_id.clone()).or_insert_with(|| HashMap::new());
-            if let Some((prev_trans, prev_vote)) =
-                votes.insert(voter.clone(), (transaction.clone(), procedure.clone()))
-            {
-                // Re-voting is allowed; new vote must be treated as the proper one,
-                // older is to be discarded.
-                if tracing::enabled!(tracing::Level::DEBUG) {
-                    debug!("Governance vote by {} for {} already registered! New: {:?}, old: {:?} from {}",
-                        voter, action_id, procedure, prev_vote, prev_trans.encode_hex::<String>()
-                    );
-                }
-            }
-        }
-        Ok(())
-    }
-
-    /// Checks whether action_id can be considered finally accepted
-    fn is_finally_accepted(
-        &self,
-        voting_state: &VotingRegistrationState,
-        action_id: &GovActionId,
-    ) -> Result<VotingOutcome> {
-        let (_epoch, proposal) = self
-            .proposals
-            .get(action_id)
-            .ok_or_else(|| anyhow!("action {} not found", action_id))?;
-        let conway_params = self.get_conway_params()?;
-        let threshold = voting_state.get_action_thresholds(proposal, conway_params)?;
-
-        let votes = self.get_actual_votes(action_id);
-        let voted = votes.majorizes(&threshold);
-        let previous_ok = match proposal.gov_action.get_previous_action_id() {
-            Some(act) => self.action_status
-                .get(&act).map(|x| x.is_accepted())
-                .unwrap_or(false),
-            None => true
-        };
-        let accepted = previous_ok && voted;
-        info!(
-            "Proposal {action_id}: votes {votes}, thresholds {threshold}, prevous_ok {previous_ok}, \
-             voted {voted}, result {accepted}"
-        );
-
-        Ok(VotingOutcome {
-            procedure: proposal.clone(),
-            votes_cast: votes,
-            votes_threshold: threshold,
-            accepted,
-        })
-    }
-
-    /// Should be called when voting is over
-    fn end_voting(&mut self, action_id: &GovActionId) -> Result<()> {
-        self.votes.remove(&action_id);
-        self.proposals.remove(&action_id);
-
-        Ok(())
-    }
-
-    /// Returns actual votes: (Pool votes, DRep votes, committee votes)
-    fn get_actual_votes(&self, action_id: &GovActionId) -> VotesCount {
-        let mut votes = VotesCount::zero();
-        if let Some(all_votes) = self.votes.get(&action_id) {
-            for voter in all_votes.keys() {
-                match voter {
-                    Voter::ConstitutionalCommitteeKey(_) => votes.committee += 1,
-                    Voter::ConstitutionalCommitteeScript(_) => votes.committee += 1,
-                    Voter::DRepKey(key) => {
-                        self.drep_stake
-                            .get(&DRepCredential::AddrKeyHash(key.clone()))
-                            .inspect(|v| votes.drep += *v);
-                    }
-                    Voter::DRepScript(script) => {
-                        self.drep_stake
-                            .get(&DRepCredential::ScriptHash(script.clone()))
-                            .inspect(|v| votes.drep += *v);
-                    }
-                    Voter::StakePoolKey(pool) => {
-                        self.spo_stake.get(pool).inspect(|ds| votes.pool += ds.live);
-                    }
-                }
-            }
-        }
-        votes
-    }
-
-    /// Checks whether action is expired at the beginning of new_epoch
-    fn is_expired(&self, new_epoch: u64, action_id: &GovActionId) -> Result<bool> {
-        info!(
-            "Checking whether {} is expired at new epoch {}",
-            action_id, new_epoch
-        );
-        let (proposal_epoch, _proposal) = self
-            .proposals
-            .get(action_id)
-            .ok_or_else(|| anyhow!("action {} not found", action_id))?;
-
-        Ok(proposal_epoch + self.get_conway_params()?.gov_action_lifetime as u64 <= new_epoch)
-    }
-
-    fn pack_as_enact_state_elem(p: &ProposalProcedure) -> Option<EnactStateElem> {
-        match &p.gov_action {
-            GovernanceAction::HardForkInitiation(_hf) => None,
-            GovernanceAction::TreasuryWithdrawals(_wt) => None,
-            GovernanceAction::Information => None,
-
-            GovernanceAction::ParameterChange(pc) => {
-                Some(EnactStateElem::Params(pc.protocol_param_update.clone()))
-            }
-            GovernanceAction::NewConstitution(nc) => {
-                Some(EnactStateElem::Constitution(nc.new_constitution.clone()))
-            }
-            GovernanceAction::UpdateCommittee(uc) => {
-                Some(EnactStateElem::Committee(uc.data.clone()))
-            }
-            GovernanceAction::NoConfidence(_) => Some(EnactStateElem::NoConfidence),
-        }
-    }
-
-    fn retrieve_withdrawal(p: &ProposalProcedure) -> Option<TreasuryWithdrawalsAction> {
-        if let GovernanceAction::TreasuryWithdrawals(ref action) = p.gov_action {
-            Some(action.clone())
-        } else {
-            None
-        }
-    }
-
-    /// Checks and updates action_id state at the start of new_epoch
-    /// If the action is accepted, returns accepted ProposalProcedure.
-    fn process_one_proposal(
-        &mut self,
-        new_epoch: u64,
-        voting_state: &VotingRegistrationState,
-        action_id: &GovActionId,
-    ) -> Result<Option<VotingOutcome>> {
-        let outcome = self.is_finally_accepted(voting_state, &action_id)?;
-        let expired = self.is_expired(new_epoch, &action_id)?;
-        if outcome.accepted || expired {
-            self.end_voting(&action_id)?;
-            info!(
-                "New epoch {new_epoch}: voting for {action_id} outcome: {}, expired: {expired}",
-                outcome.accepted
-            );
-            return Ok(Some(outcome));
-        }
-
-        Ok(None)
-    }
-
-    fn gov_action_id_to_string(action_id: &GovActionId) -> String {
-        format!(
-            "transaction: {}, action_index: {}",
-            hex::encode(action_id.transaction_id),
-            action_id.action_index
-        )
-    }
-
-    fn get_action_name(action: &GovernanceAction) -> &str {
-        match action {
-            GovernanceAction::ParameterChange(_) => "ParameterChange",
-            GovernanceAction::HardForkInitiation(_) => "HardForkInitiation",
-            GovernanceAction::TreasuryWithdrawals(_) => "TreasuryWithdrawals",
-            GovernanceAction::NoConfidence(_) => "NoConfidence",
-            GovernanceAction::UpdateCommittee(_) => "UpdateCommittee",
-            GovernanceAction::NewConstitution(_) => "NewConstitution",
-            GovernanceAction::Information => "Information"
-        }
-    }
-
-    /// Function dumps information about completed (expired, ratified, enacted) governance 
-    /// actions in format, close to that of `gov_action_proposal` from `sqldb`.
-    fn print_outcome_to_verify(&self, outcome: &Vec<GovernanceOutcome>) -> Result<()> {
-        let out_file_name = match &self.verification_output_file {
-            Some(o) => o,
-            None => return Ok(())
-        };
-
-        if outcome.is_empty() {
-            return Ok(());
-        }
-
-        let mut out_file = match OpenOptions::new().append(true).open(out_file_name.clone()) {
-            Ok(res) => res,
-            Err(e) => bail!("Cannot open verification output {out_file_name} for writing: {e}")
-        };
-
-        // id,tx_id,index,prev_gov_action_proposal,deposit,return_address,expiration,
-        // voting_anchor_id,type,description,param_proposal,ratified_epoch,enacted_epoch,
-        // dropped_epoch,expired_epoch
-        for elem in outcome.iter() {
-            let prev_action = match &elem.voting.procedure.gov_action.get_previous_action_id()
-            {
-                Some(act) => Self::gov_action_id_to_string(act),
-                None => "".to_owned()
-            };
-
-            let action_status = self.action_status.get(
-                &elem.voting.procedure.gov_action_id
-            ).ok_or_else(|| anyhow!("Cannot get action status for {}",
-                &elem.voting.procedure.gov_action_id
-            ))?;
-
-            let deposit = &elem.voting.procedure.deposit;
-            let reward = hex::encode(elem.voting.procedure.reward_account.to_vec());
-            let ratification_info = if elem.voting.accepted {
-                format!(
-                    "{:?},{:?}", action_status.ratification_epoch, action_status.enactment_epoch
-                )
-            }
-            else {
-                ",".to_owned()
-            };
-            let ptype = Self::get_action_name(&elem.voting.procedure.gov_action);
-            let expired = action_status.expiration_epoch.to_string();
-            let res = format!(
-                ",,,{prev_action},{deposit},{reward},,,{ptype},,,{ratification_info},,{expired}\n"
-            );
-            if let Err(e) = out_file.write(&res.as_bytes()) {
-                error!(
-                    "Cannot write 'res' to verification output {out_file_name} for writing: {e}"
-                );
-            }
-        }
-
-        Ok(())
-    }
-
-    fn finalize_conway_voting(
-        &mut self,
-        new_block: &BlockInfo,
-        voting_state: &VotingRegistrationState,
-    ) -> Result<Vec<GovernanceOutcome>> {
-        let mut outcome = Vec::<GovernanceOutcome>::new();
-        let actions = self.proposals.keys().map(|a| a.clone()).collect::<Vec<_>>();
-
-        for action_id in actions.iter() {
-            info!("Epoch {} started: processing action {}", new_block.epoch, action_id);
-            match self.process_one_proposal(new_block.epoch, &voting_state, &action_id) {
-                Err(e) => error!("Error processing governance {action_id}: {e}"),
-                Ok(None) => (),
-                Ok(Some(out)) if out.accepted => {
-                    let mut action_to_perform = GovernanceOutcomeVariant::NoAction;
-
-                    if let Some(elem) = Self::pack_as_enact_state_elem(&out.procedure) {
-                        action_to_perform = GovernanceOutcomeVariant::EnactStateElem(elem);
-                    } else if let Some(wt) = Self::retrieve_withdrawal(&out.procedure) {
-                        action_to_perform = GovernanceOutcomeVariant::TreasuryWithdrawal(wt);
-                    }
-
-                    outcome.push(GovernanceOutcome {
-                        voting: out,
-                        action_to_perform,
-                    })
-                }
-                Ok(Some(out)) => outcome.push(GovernanceOutcome {
-                    voting: out,
-                    action_to_perform: GovernanceOutcomeVariant::NoAction,
-                }),
-            }
-        }
-
-        Ok(outcome)
-    }
 
     fn recalculate_voting_state(&self) -> Result<VotingRegistrationState> {
         let drep_stake = self.drep_stake.iter().map(|(_dr, lov)| lov).sum();
 
-        let committee_usize = self.get_conway_params()?.committee.members.len();
+        let committee_usize = self.conway_voting.get_conway_params()?.committee.members.len();
         let committee = committee_usize.try_into().or_else(|e| {
             Err(anyhow!(
                 "Commitee size: conversion usize -> u64 failed, {e}"
@@ -524,7 +171,10 @@ impl State {
 
         if self.current_era >= Era::Conway {
             let voting_state = self.recalculate_voting_state()?;
-            let outcome = self.finalize_conway_voting(&new_block, &voting_state)?;
+            let ratified = self.conway_voting.finalize_conway_voting(
+                &new_block, &voting_state, &self.drep_stake, &self.spo_stake
+            )?;
+            let outcome = self.conway_voting.put_outcomes_to_queue(new_block.epoch, ratified)?;
             let acc = outcome.iter().filter(|oc| oc.voting.accepted).count();
 
             info!(
@@ -534,43 +184,21 @@ impl State {
                 outcome.len()
             );
 
-            self.log_voting_stats();
+            self.conway_voting.log_conway_voting_stats();
             info!("Conway voting: new epoch {}, outcomes: {outcome:?}", new_block.epoch);
             output.conway_outcomes = outcome;
         }
 
-        self.print_outcome_to_verify(&output.conway_outcomes)?;
-        return Ok(output);
+        self.conway_voting.print_outcome_to_verify(&output.conway_outcomes)?;
+        Ok(output)
     }
 
     fn log_stats(&self) {
-        info!("props: {}, props_with_id: {}, votes: {}, stored proposal procedures: {}, drep stake msgs (size): {} ({})",
-            self.proposal_count, self.action_proposal_count, self.votes_count, self.proposals.len(),
+        info!("{}, {}, drep stake msgs (size): {} ({})",
+            self.alonzo_babbage_voting.get_stats(),
+            self.conway_voting.get_stats(),
             self.drep_stake_messages_count, self.drep_stake.len(),
         );
-    }
-
-    fn log_voting_stats(&self) {
-        info!("*** Current voting stats: ***");
-        let mut proposal_procedures = self.proposals.keys().cloned().collect::<HashSet<GovActionId>>();
-        for (action_id, voting_procedure) in self.votes.iter() {
-            let proposal = match self.proposals.get(action_id) {
-                None => " (absent) ".to_string(),
-                Some(p) => {
-                    proposal_procedures.remove(action_id);
-                    format!(" {p:?} ")
-                }
-            };
-            info!(
-                "{action_id}: {proposal} => {voting_procedure:?}",
-            )
-        }
-
-        if !proposal_procedures.is_empty() {
-            let pp = proposal_procedures.into_iter().map(|x| format!("{x},")).collect::<String>();
-            info!("Proposal procedures without 'votes' records: [{}]", pp);
-        }
-        info!("*** End of voting stats ***");
     }
 
     pub async fn send(&self, block: &BlockInfo, message: GovernanceOutcomesMessage) -> Result<()> {
@@ -593,12 +221,12 @@ impl State {
 
     /// Get list of actual voting proposals
     pub fn list_proposals(&self) -> Vec<GovActionId> {
-        self.proposals.keys().cloned().collect()
+        self.conway_voting.proposals.keys().cloned().collect()
     }
 
     /// Get details for a specific proposal
     pub fn get_proposal(&self, id: &GovActionId) -> Option<ProposalProcedure> {
-        self.proposals.get(id).map(|(_epoch, prop)| prop.clone())
+        self.conway_voting.proposals.get(id).map(|(_epoch, prop)| prop.clone())
     }
 
     /// Get list of votes for a specific proposal
@@ -606,7 +234,7 @@ impl State {
         &self,
         proposal_id: &GovActionId,
     ) -> Result<HashMap<Voter, (TxHash, VotingProcedure)>> {
-        match self.votes.get(proposal_id) {
+        match self.conway_voting.votes.get(proposal_id) {
             Some(all_votes) => Ok(all_votes.clone()),
             None => Err(anyhow::anyhow!(
                 "Governance action: {:?} not found",
